Cambios y mejoras minimas

+++ app/Http/Controllers/ContratoController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use App\Models\Contrato;
 use App\Models\Tipocontrato;
 use App\Models\Personanatural;
+use App\Models\Proceso;
 use App\Models\Pago;
 use App\User;
 use App\Http\Requests\ContratoFormRequest;
@@ -29,19 +30,22 @@ class ContratoController extends Controller
     {
         $palabrasbuscar = explode(" ",$request->post('buscar'));
         $contratos = Contrato::orderBy('id', 'ASC')
-                        ->select('contrato.*', 'tipocontrato.descripcion AS tipocontrato')
+                        ->select('contrato.*', 'tipocontrato.descripcion AS tipocontrato',
+                                    'personanatural.numerodocumento', 'proceso.numero AS numeroproceso')
                         ->selectRaw('CONCAT(personanatural.nombres, " ", personanatural.apellidopaterno, " ", personanatural.apellidomaterno) AS nombrecompleto')
                         ->join('tipocontrato','tipocontrato_id','=','tipocontrato.id')
-                        ->join('personanatural','personanatural_id','=','personanatural.id');
+                        ->join('personanatural','personanatural_id','=','personanatural.id')
+                        ->join('proceso','proceso_id','=','proceso.id');
         $emptypalabrasbuscar = array_filter($palabrasbuscar);
         if (!empty($emptypalabrasbuscar)){
-            $columnas = ['valor', 'numero', 'tipocontrato.descripcion', 
-            'personanatural.nombres', 'personanatural.apellidopaterno', 'personanatural.apellidomaterno'];
+            $columnas = ['contrato.valor', 'contrato.numero', 'tipocontrato.descripcion', 
+                'personanatural.nombres', 'personanatural.apellidopaterno', 
+                'personanatural.apellidomaterno', 'personanatural.numerodocumento', 'proceso.numero'];
             $Contratos['Contratos'] = $contratos
             ->whereOrSearch($palabrasbuscar, $columnas);
                     
             return view('contrato.index', $Contratos)
-            ->with('success','Busqueda realizada');
+            ->with('success',['Busqueda realizada']);
         }else{
             $Contratos['Contratos'] = $contratos->paginate(10);
             return view('contrato.index', $Contratos);
@@ -57,7 +61,8 @@ class ContratoController extends Controller
     {
         $Tipocontratos = Tipocontrato::select('id', 'descripcion')->get();
         $Personasnaturales = Personanatural::select('id', 'nombres', 'apellidopaterno', 'apellidomaterno')->get();
-        return view('contrato.create', compact('Tipocontratos', 'Personasnaturales'));
+        $Procesos = Proceso::select('id', 'numero')->get();
+        return view('contrato.create', compact('Tipocontratos', 'Personasnaturales', 'Procesos'));
     }
 
     /**
@@ -75,8 +80,8 @@ class ContratoController extends Controller
         $path = $file->storeAs(self::PATH. $request->personanatural_id, $fileName);
         if (Storage::exists($path)){
             return redirect()->route('contrato.index')
-                ->with('success',['data1' => 'Actuación del proceso almacenado completamente', 
-                'data2' => 'Archivo de actuación del proceso almacenado completamente']);
+                ->with('success',['Actuación del proceso almacenado completamente', 
+                'Archivo de actuación del proceso almacenado completamente']);
         }else{
             return redirect()->route('contrato.index')
                 ->with('success',['Error no se escribio archivo en disco']);
@@ -91,10 +96,11 @@ class ContratoController extends Controller
      */
     public function show(Contrato $contrato)
     {
-        $auditoria = User::findOrFail($contrato)->first();
-        $tipocontrato = Tipocontrato::select('descripcion')->find($contrato->tipocontrato_id);
-        $personanatural = Personanatural::select('nombres', 'apellidopaterno', 'apellidomaterno')->find($contrato->personanatural_id);
-        return view('contrato.show', compact('tipocontrato', 'personanatural', 'contrato', 'auditoria'));
+        $auditoria = User::findOrFail($contrato->users_id);
+        $tipocontrato = Tipocontrato::select('descripcion')->findOrFail($contrato->tipocontrato_id);
+        $personanatural = Personanatural::select('nombres', 'apellidopaterno', 'apellidomaterno')->findOrFail($contrato->personanatural_id);
+        $procesos = Proceso::select('id', 'numero')->get();
+        return view('contrato.show', compact('tipocontrato', 'personanatural', 'contrato', 'auditoria', 'procesos'));
     }
 
     /**
@@ -107,7 +113,8 @@ class ContratoController extends Controller
     {
         $Tipocontratos = Tipocontrato::select('id', 'descripcion')->get();
         $Personasnaturales = Personanatural::select('id', 'nombres', 'apellidopaterno', 'apellidomaterno')->get();
-        return view('contrato.edit', compact('contrato', 'Tipocontratos', 'Personasnaturales'));
+        $Procesos = Proceso::select('id', 'numero')->get();
+        return view('contrato.edit', compact('contrato', 'Tipocontratos', 'Personasnaturales', 'Procesos'));
     }
 
     /**
@@ -131,9 +138,9 @@ class ContratoController extends Controller
         $result = $this->deleteFile($personanatural_anterior, $nombrearchivo_anterior);
         if (Storage::exists($path)){
             return redirect()->route('contrato.index')
-            ->with('success',['data1' => 'Contrato actualizado', 
-            'data2' => 'Archivo previo borrado completamente: '. $nombrearchivo_anterior,
-            'data3' => 'Se carga nuevo archivo: '. $fileName]);
+            ->with('success',['Contrato actualizado', 
+            'Archivo previo borrado completamente: '. $nombrearchivo_anterior,
+            'Se carga nuevo archivo: '. $fileName]);
         }else{
             return redirect()->route('contrato.edit')
             ->with('success',['Error no se escribio archivo en disco']);
@@ -146,31 +153,41 @@ class ContratoController extends Controller
      * @param  \App\Contrato  $contrato
      * @return \Illuminate\Http\Response
      */
-    public function destroy(Contrato $contrato)
+    public function destroy($id)
     {
-        //$actuacionproceso =  Actuacionproceso::find($id);
-        $valida = Pago::where('contrato_id', '=', $contrato->id)->get();
+        $valida = Pago::where('contrato_id', '=', $id)->get();
         if ($valida->isEmpty()) {
-            $contrato->delete();
-            $result = $this->deleteFile($contrato->personanatural_id, $contrato->nombrearchivo);
-            return redirect()->route('contrato.index')->with('success',
-            [
-                'data1' => 'Registro borrado completamente',
-                'data2' => 'Archivo borrado completamente' . $result
-            ]);
-        }else{
-            return redirect()->route('contrato.index')
-            ->withErrors(['No se puede borrar el contrato', 'El contrato tiene pagos asociados']);
+            $contrato = Contrato::findOrFail($id);
+            $result1 = $contrato->delete();
+            $result2 = $this->deleteFile($contrato->personanatural_id, $contrato->nombrearchivo);
+            if($result1 && $result2)
+                return redirect()->route('contrato.index')->with('success',
+                    ['Registro borrado completamente','Archivo borrado completamente'
+                    ]);
         }
+        
+        return redirect()->route('contrato.index')
+            ->withErrors(['No se puede borrar el contrato', 'El contrato tiene pagos asociados']);
+        
     }
 
     public function downloadFile($id, $name)
     {
-        return Storage::download(self::PATH. $id.'/'.$name);
+        $file = Storage::exists(self::PATH. $id.'/'. $name);
+        if ($file){
+            return Storage::download(self::PATH. $id.'/'.$name);
+        }
+        return redirect()->route('contrato.index')->withErrors(['No se encuentra el archivo: '. $name]);
     }
 
     public function deleteFile($id, $name)
     {
-        return Storage::delete(self::PATH. '/'. $id.'/'.$name);
+        $file = Storage::exists(self::PATH. $id.'/'. $name);
+        if ($file){
+            return Storage::delete(self::PATH. '/'. $id.'/'.$name); 
+        }
+            
+        return redirect()->route('contrato.index')->withErrors(['No se encuentra el archivo: '. $name]);
     }
 }
+
